+++ tests/test_core/test_layout.py
@@ -8,9 +8,10 @@ from weakref import ref as weakref
 import pytest
 
 import idom
+from idom import html
 from idom.config import IDOM_DEBUG_MODE
 from idom.core.dispatcher import render_json_patch
-from idom.core.hooks import use_effect
+from idom.core.hooks import use_effect, use_state
 from idom.core.layout import LayoutEvent
 from idom.testing import (
     HookCatcher,
@@ -838,3 +839,87 @@ async def test_changing_key_of_parent_element_unmounts_children():
             root_hook.latest.schedule_render()
             await layout.render()
             assert last_state != state.current
+
+
+async def test_switching_node_type_with_event_handlers():
+    toggle_type = idom.Ref()
+    element_static_handler = StaticEventHandler()
+    component_static_handler = StaticEventHandler()
+
+    @idom.component
+    def Root():
+        toggle, toggle_type.current = use_toggle(True)
+        handler = element_static_handler.use(lambda: None)
+        if toggle:
+            return html.div(html.button({"onEvent": handler}))
+        else:
+            return html.div(SomeComponent())
+
+    @idom.component
+    def SomeComponent():
+        handler = component_static_handler.use(lambda: None)
+        return html.button({"onAnotherEvent": handler})
+
+    with idom.Layout(Root()) as layout:
+        await layout.render()
+
+        assert element_static_handler.target in layout._event_handlers
+        assert component_static_handler.target not in layout._event_handlers
+
+        toggle_type.current()
+        await layout.render()
+
+        assert element_static_handler.target not in layout._event_handlers
+        assert component_static_handler.target in layout._event_handlers
+
+        toggle_type.current()
+        await layout.render()
+
+        assert element_static_handler.target in layout._event_handlers
+        assert component_static_handler.target not in layout._event_handlers
+
+
+async def test_switching_component_definition():
+    toggle_component = idom.Ref()
+    first_used_state = idom.Ref(None)
+    second_used_state = idom.Ref(None)
+
+    @idom.component
+    def Root():
+        toggle, toggle_component.current = use_toggle(True)
+        if toggle:
+            return FirstComponent()
+        else:
+            return SecondComponent()
+
+    @idom.component
+    def FirstComponent():
+        first_used_state.current = use_state("first")[0]
+        # reset state after unmount
+        use_effect(lambda: lambda: first_used_state.set_current(None))
+        return html.div()
+
+    @idom.component
+    def SecondComponent():
+        second_used_state.current = use_state("second")[0]
+        # reset state after unmount
+        use_effect(lambda: lambda: second_used_state.set_current(None))
+        return html.div()
+
+    with idom.Layout(Root()) as layout:
+        await layout.render()
+
+        assert first_used_state.current == "first"
+        assert second_used_state.current is None
+
+        toggle_component.current()
+        await layout.render()
+
+        assert first_used_state.current is None
+        assert second_used_state.current == "second"
+
+        toggle_component.current()
+        await layout.render()
+
+        assert first_used_state.current == "first"
+        assert second_used_state.current is None